chat-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLModel==0.0.8
asyncpg==0.29.0
sentence-transformers==2.3.1
pgvector==0.2.4
python-dotenv==1.0.0
google-generativeai==0.4.0
slowapi==0.1.5
prometheus-fastapi-instrumentator==6.1.0
pytest==7.4.3
pytest-asyncio==0.23.2 